Add missing "end of tomorrow" option to UI dropdown
The End dropdown in vrm-api.html is missing the "end of tomorrow" (eot) option that is already supported by the parser in stats-parameters.js.

This commit adds the missing "end of tomorrow" option to the End dropdown, enabling the use case: Start "beginning of tomorrow" + End "end of tomorrow" for retrieving complete next-day forecasts.