fix: add Dynamic ESS schedule endpoint
The scheduling endpoint now actually returns schedules that can be inserted into the dbus scheduler.

Note that the stats Dynamic ESS endpoint returned. The stats endpoint is for creating graphs (price, planning), not for actually scheduling. In
order to make most of that, use the _Transform price schedule to second output_ checkbox.

See the examples for more info.